fix(html): character references manquant dans les attributs des éléments #18
Fix #18

+++ html/named_characters/mod.rs
@@ -26,7 +26,7 @@ pub struct NamedCharacterReferences(NamedCharacterReferencesEntities);
 #[derive(Debug)]
 #[derive(Deserialize)]
 pub struct NamedCharacterReferenceEntity {
-    pub codepoints: Vec<usize>,
+    pub codepoints: Vec<u32>,
     pub characters: String,
 }
